requirements.txt
Django>=5.0
balldontlie>=0.1.6
whitenoise[brotli]>=6.6
gunicorn>=21.2
python-dotenv>=1.0.0
psycopg2
requests>=2.31.0
beautifulsoup4>=4.12.0
Pillow>=12.0.0
markdown2>=2.4.0
django-allauth>=0.57.0
PyJWT>=2.8.0
cryptography>=41.0.0
